Refactor: backend architecture based on database
