+++ src/backend/models.py
@@ -1,0 +1,21 @@
+from sqlalchemy.orm import declarative_base
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from datetime import datetime
+from sqlalchemy.orm import relationship
+
+Base = declarative_base()
+
+class Experiment(Base):
+    __tablename__ = "experiments"
+    id = Column(String, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    created_at = Column(DateTime, default=datetime.now())
+    user = relationship("User", back_populates="experiments")
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String, nullable=False, unique=True)
+    password = Column(String, nullable=False)
+    created_at = Column(DateTime, default=datetime.now())
+    experiments = relationship("Experiment", back_populates="user", cascade="all, delete-orphan")
